refactor: 上移 `commands/triggers/` 目录
