+++ src/features/command-triggers/script-event.ts
