--- src/features/commands/index.ts
@@ -1,1 +1,0 @@
-export * from './triggers';
